Video upload now working
Problem was related to the Azure url. It contains query params so we have to remove them first in order to get original_file name correctly. Since we were not getting the original_filename with the .mp4 extension, ffmpeg did not know what to do with it and threw an error.