Fixed issues with hotkeys
Seemed that the real problem was not so much with hot keys but with the selected prop of the instances. When we changed frames the watcher would always reset to false the instance select and that would cause some of the hotkeys for related to selection to stop working (like copy and paste). 

Please let me know if you are aware of any other hotkeys malfunctioning so I can dive deeper into them here.