1) Loading context 2)name video but for both images and video
For loading context, if we do the block here, then all of the ghostloaders and other stuff for each individual section don't show as well. I see value in this for when we are doing tiling with multiple annotation_cores, but let's address that use case then as it feels like a very different context. Suggesting we leave hook but otherwise leave it up to individual components to handle loading