feat: Extract global frame numbers for frame packet map in video split
Idea: If my math is now wrong, if I multiply the offset in seconds times the fps, it will give us the global frame number. After that we can condition on each video to see if its a splitted video and then use the global frame number to get the instance list.