feat: handle case of issue_panel by hiding
It seems unexpected that selecting a ghost instance in the issue context creates an attachment for it. For now I think it’s ok to for the “promotion” of the instance before an issue is created.
Not sure if this a good long term thing, but it gives a point to think about this type of interaction