feat: switched file mapping to questions format.
Now we ask the user for each field making one questions per field only. This is to improve ease of user and make the user focus on only one field at the time. I also added more details on the text for a better explanation of what each field does.