add further guards here - more root cause of JS scripts out of order load
The logic for determine when we change is not great.
Right now it mixes both the editing and usage, meaning we have a prop for cases where it's externally loaded, but then a somewhat circular reference, causing it to fire twice since the prop watcher changes.

Ideally, we probably need to more separation of concerns of these components.