fix: fix video playing error
The error was the combination of a couple things. First, it was a mistake to add the go_to_keyframe on mount since it eventually sets video.src. According to docs, doing video.src = something is the same as loading the video. That's why the video play error was happening. Second, I did not saw that annoation_core did some preloading on the created() method too and not only mounted. So I added that logic to make the first frame appear when initial loading,