Dataset Explorer (#90)
* feat: created file preview component

* wip: adding instance list to file preview.

* wip: display instance list

* wip: resizing canvas and instances

* fix: indents

* wip: resize instance list

* wip: instance_list resizing

* wip: resizing instance list

* wip: instance list resizing now working.

* feat: images now load correctly on first paint

* wip: adding model run selector component

* feat: added model run list endpoint.

* wip: model run selector

* feat: model run filtering working for base selection

pending compare to instance lists.

* feat: filter instance list buy run id and show/hide grountruth

* feat: clicking image to view instances now working

* wip: adding model run concepts into annotation core.

* feat: add filter by id to model run list

* feat: adding accordeon grouping for model runs

Pending color mapping

* feat: Add show/hide model runs

Now we can show/hide the entire group of the model run on the instance_list detail component.

* feat: add ground truth list

* wip: adding video to drawable canvas

* wip: video drawable canvas

* feat: video player now working in preview

Seems that the problem was that the refresh prop needed to be updated since there is a wrapper parent component for this case.

* feat: adjusted canvas scaling when playing the video

The video was being played but the scale was a bit off. Now the scale is working fine.

* wip: displays instance on video preview

Pending adding the model run filters and disabling query params on video component

* feat: filters now working for video previews

* feat: display accordeon component for video files.

* wip: remove query params matching and improve layout on explorer

* feat: improved scrollbar and first video paint

* wip: improve scrolling. 

Still need to add the infinite scroller and the virtual scroller to improve performance.

* feat: added infinite scrolling

Removed libraries because they were not quite working as expected. Implemented the logic manually

* fix: fix refresh problem when changing dirs.

* fix: fix video playing error

The error was the combination of a couple things. First, it was a mistake to add the go_to_keyframe on mount since it eventually sets video.src. According to docs, doing video.src = something is the same as loading the video. That's why the video play error was happening. Second, I did not saw that annoation_core did some preloading on the created() method too and not only mounted. So I added that logic to make the first frame appear when initial loading,

* feat: added machine_made concept

* fix: fix height issue when changing from data explorer tab to file tab

* cleanup

* fix: fix e2e tests

* fix: fix preview so it works based on local width and height

* fix: update accordeon properly when removing model runs

Co-authored-by: Pablo <pjestradac@gmail.com>