fix: in review state has no way to complete.
I'm changing this to completed once they click complete again as it's the more straightforward way of going about it. We can discuss the ideal way to manage the review states in the future or even discuss if those states are necessary at all now that we have the issues system.