fix: make process media thread safe with locks (#126)
* fix: make process media thread safe with locks

The context is that since queues are fetched by multiple threads at the same time. We can get a chance were 2 threads process the same item. We add a thread lock to both the frame and the video queue to prevent this from happening as an exception raised at this level can halt the entire system and require a restart.

* fix: remove exception format

Sometimes this function can fail so removing it for now as this is the final place to catch an exception in the execution tree.

Co-authored-by: Pablo <pjestradac@gmail.com>