feat: improve pagination on file list
The idea is to remove the boolean request_next page and just build the pagination mechanism based on the page number and the limit.

This is because now the pagination is going to be supported in the SDK and we need this working before we can implement the new SDK features.