fix: update frame buffer dict if frame was changed during save (#213)
This happens when the user does not save the file and changes the frame. Since the update algorithm always updated the instance_list, and the frame number has changed. The instance_list now corresponds to another frame, so now we pass the requests video data to check which was the frame that was saved and update the buffer dict accordingly.

Co-authored-by: Pablo <pjestradac@gmail.com>