fix: handle Forbidden exceptions correctly (#299)
* fix: handle Forbidden exceptions correctly

The exception handle we added was catching all types of exceptions. But we need to make a special case for the Forbidden exception since it's a different status code. 

From inspecting the code the are no other types of exceptions other than IOError, but that one should be handled as a 500. I think the general route should be to default to 500 and handle special types of exceptions with other handlers in this file.

* fix: remove error handlers in main.py

they have been moved to shared/error_handlers.py

Co-authored-by: Pablo <pjestradac@gmail.com>