fix: not rendering
before in mounted() on hot refresh, or sometimes just on general, it wouldn't seem to load as expected. Removing aysnc from mounted appeared to fix it so it appears to be some kind of conflict with the async/await pattern maybe.
So as to not disrupt set_bg() I just moved this to created which appears to be fine