fix: better visbility to frame update errors
When an update to a video failed on one frame, the error was not being propagated to the parent and resulted in the video on a processing state forever. We now capture that and set the parent file to failing state.

Also there was an exception being raised when trying to update a sequence with instance with instance.frame_number = None

Finally added some extra validations to prevent None type instances being processed by the walrus.