Add rest of specs to be processed
We need to have a list of these strings allowed somewhere anyway. In that context, instead of having it be a default value, it seems more clear to simply repeat the de facto default of `ui_schema_button_spec_list`. `ui_schema_logo_spec_list` provides an example of having a different specification for a different element.