Add: Default Navigation Wizard
The main idea here is that for default cases we have a consistent navigation experience.
Each of the buttons implements a visible check with a matching named slot. So if for any reason we need to override the button we can easily do that and still keep the rest of the buttons and general spacing assumptions.

Also since disabled seems to be a common flag that is built in. We can build in other flags to avoid slot use too