Fix: Bounds (mostly)
1) Map the bounds (0 -> max size) to world coordinates. Following the example of mouse transform that uses ctx matrix.
2) Generate proposed point based on the allowed bands and the movement

The main insight here is that the mouse position shouldn't effect if we can pan or not. For example, becuase I may drag the mouse on one side of screen, but the question is really if that action will cause the edge to go out of bounds.

This does not fix artifacts on zoom out near edges. There are also a few cases this doesn't quite solve I will work on