Fix: Bug, handle if the component is created After data available
If used in a menu, then this component may get created AFTER the data is already loaded.  this meant it would fail to show. so handle the on creation here.

This was (upgraded?) to use v-model and in the process this regression was caused. the info_multiple just checks the keys directly on v-if so appears to not be effected by this