Improve: Set the value directly - fix bug auto set to false
This eliminates the need to get the current config becuase it only sets the desired key.
This fixes a bug where setting the desired instance types would auto set the `visible` to false, which was becuase passing this as a prop doesn't work with vuex it appears.

Also move to the get_ui_schema() for similar reason, it will work initially but won't update from vuex

For the concrete methods that are different for every case, for now I'd prefer to leave it on the component directly. L165 does everything needed without an extra event. I think in this exact case I find it confusing becuase it's emiting locally, only to then emit globally to vuex. I'd rather just commit directly to vuex I think in this case?