Opinion: On Names, stating it as a question
We already state the literal thing like select as the sub header or label on selector. So instead of repeating it this phrases it in a different way that's more conversational. It still leaves the regular select message.

I think at most we should have 3 lines here. 1 as the header, 2 as the sub header, and then 3rd the literal select if required.