Actions And Accounts Features From EE (#400)
* wip: migrate methods

* wip: adding org and actions

* fix: db imports

* wip: add marketing pages

* fix: e2e tests

* fix: alembic migration

* wip: adding missing menus

* wip: adding actions frontend

* reorder tests

* wip: adding actions, remove brain

* wip: adding actions

* wip: adding actions migrations and account migrations to opencore

* wip: add stripe billing

* feat: add kubernetes deploy github action

* fix: add stripe requirement

* fix: add newrelic docker images

* feat: add NEW_RELIC_LICENSE_KEY

* feat: add new relic.ini

* fix: switch to opencore on helm chart

* remove name

* Fixes from merge conflict

* Bugfix L369 - If logged out then user_projects_list may have no length

* Fix Console Error - missing component

* Maintain text name

Otherwise if the user chooses a caps sensitive name it gets all caps

* Remove old and identify legacy

We aren't using the usage cycle at  the moment because we doing unlimited per user billing. The run_stripe is a manual bad file and should just be deleted.

* Clean

icon is the default mode

* this was an old thing

* add the check_export_billing() for diffgram.com

at the moment if settings.ALLOW_STRIPE_BILLING is False it just returns. This may be a good place to put in checks for say being at 20 users and stuff like that (in open source context).

* Make Reply To work in this context

* Remove old files

All of these have been replaced by the new unbounce marketing pages, or docs etc. Some of the high level names are actually fine, but the content is out of date or wrong etc. so seems easier to just straight delete and can recreate name later if needed

* Remove old contact and mark legacy

the examples and call out are used in another component. Both are extremely out of date but parent component is also hidden. Maybe worth having as a quick starting point to revisit in future

* Align contact and try buttons

Remove old pages from router
Fix bug with new user route

* Add undefined guard

* Bug fix small

* fix: unit tests

* fix: dashboard error

* fix: small bugfixes

fix some small bugs while doing general usage of the platform

Co-authored-by: Pablo <pjestradac@gmail.com>
Co-authored-by: Anthony Sarkis <18080164+anthony-sarkis@users.noreply.github.com>