Improve sequence list (#489)
* remove legacy code

1) this is legacy -> default_sequences_to_single_frame
2) Clean comments

* Align new regen function and refactor

1) regenerate_keyframes_for_whole_series()
now makes more sense for case of actually wanting to regen the whole list
2) Clarifies the use case of updating from a single instance
3) Refactors is_keyframe_missing_from_list()
Although as a note here given the `not in list` check, perhaps a try/catch block with the binary search would be more efficient

* add org case and wordsmith

1) org case is intent is to not show message to existing orgs
may need to review if we are assigned default ids
2) Wordsmith to save space on top bar

* Fix branching case

Assuming branching case was not intentional
restores linear order

* clean: remove legacy code  ci skip

hide print statement

* Fix: SQL issue

psycopg2.errors.InvalidTextRepresentation: invalid input syntax for integer: ""
LINE 1: ...e_updated) VALUES (2, true,  -1, false,  -1,  -1, '',  -1,  ...

* Clean and fix missing self

* Fix: Extract keyframe cache to be more composable

1) Split update into declarative add/remove
2) Split the keyframe cache from other sequence work to make it closer to the actual decision point

* Add: skip extra logging for annotation update params

This was an extra logging thing. perhaps could be optional in future.
It information dumps way too much content to be  clearly useful as far as I can tell.

* Improve: Fix logic for when we ignore deletion

* clean ci skip

* Bugfixs: Adding all to all keyframes on front end

1) Now conditions on the action_type of the instance (eg added_instances, or 'deleted')  This replaces refresh_sequence_frame_list()

2) Added deletion context on this front end case
`remove_frame_number_from_sequence`

The benefit of leaning on backend here is that it helps ensure alignment, but the downside is that it's a little clunky. We could instead do this before save

* Clean: ci skip

* Go back to hard error on outdated list - can revert if needed

1) Because this way reproduced further issues with keyframes so far
2) Context of that other ticket where we assumed this was firing

* Switch to current_frame_cache to avoid issues on frame change

otherwise it injects the wrong frame

* Adjust context now that hard check is re-enabled

* Revert "Adjust context now that hard check is re-enabled"

This reverts commit 99fbb01d2d52e4fe7e41b9c9f0e9f35b3278664c.

* Revert "Go back to hard error on outdated list - can revert if needed"

This reverts commit cd2122a9bb38519a0dd31740031494e88369d8fe.