Fix Concurrent Saving Frames Issues (#491)
* fix: add pending save to each instance iteration during paste

This helps preventing the UI to change frame without saving ids. It will keep the pending_changes variable. Still need to implement a way to save all left out pending frames even if the current_frame is not the same.

* fix: merge conflicts

* refactor: renaming function to add_instance_to_file and adding frame 

Moving toward function relying more on the frame number as a parameter, pending UpdateCommands refactor

* fix: added blocks for preventing rapid copy paste.

We now wait for frames to be fully loaded before allowing any type of paste operation.

* fix: unit tests

* fix: unit tests

* fix: fixes in polygon creation and remove frame number from image case

* fix: cuboid ghost instances line dash

* fix: remove this.instance_list cache concept

We now pass the instance list as parameter to keep it in function context.

* fix: e2e test

* fix: autosave on image case

* fix: e2e tests

* test: add unit tests

* test: unit tests

* test: finish unit tests

* fix: add save when changing frames via slider. 

To guarantee saving before resetting the frame buffer dict

* fix: assertion

* wip: refactoring sequence update to support multiple sequences

* fix: add auto increase sequence functionality

* fix: deletion on sequence that have not been saved.

* wip: cleanup

* fix: tests

* fix: e2e tests

Co-authored-by: Pablo <pjestradac@gmail.com>