fix: remove the @Input event since it causes double changes of
In the context of the slider, this event was causing the frame to "regress" to the previous number because it would be triggered every time the frame number changed. I don't see a use of this so removing it to concurrent changes of the frame number thanks to this event handler.