Default file level options - WIP (#289)
* Add is_global concept WIP

* Fix typo and simplify

* Create alembic_2021_09_14_12_01_e1739d7f895f_add_global_attributes.py

* Update attribute_template_group.py

* opinon, the ID is more of a super admin thing

but the instance counts could be for all
another option would be to use label_settings to enable or disable IDs
in context of non-expert users like Drs etc IDs maybe not what we want

* basic idea of initializing global instances

* add global UI rendering

remove unused import (axios)

* transmit global instances

* transmit updates to parent

* add value here to pass type checking on save

for now it seems better to do this than reconsider the label_file_id for the defualt case. but this is a stop gap probably

* allow globals to save

small refactor to make deduct_spatial_coordinates() easier to read
and easier to think about with global

* handle global case

* add assumed data. create throw to handle this specific case

* better to do this here

We still set it to None later to avoid foreign key violations
Not saying this is a good way to do it!
I also didn't want to mutate the spec list, to ensure the normal instances are checked properly

* opinon, I think the text here was confusing more than helpling

I think I added this a while ago when we were trying to "Declare" each part of the interface.
As we keep improving the design the text here starts to confuse more than it helps. This is valuable space, especially as more attributes, globals etc

* move into expansion panels - a work in progress

* Clean further

* Merge fix, align with master

* Fix: Out of order DB revision

* Handle a case with resetting the name

Context that we want a local copy for being able to edit the name
but we aren't getting the job data from this component
Not saying this is ideal but should work for now

* Fix merge

and clean old comments

* fix merge and clean

* wip: wizard

* feat: refactor relations to instance type

* fix: alembic migrations

* test: add unit tests

* feat: add wizard step for global attributes and improve layouts

* feat: save global attributes and display them

* fix: alembic migrations

* feat: allow global attributes in video.

* fix: unit tests

* fix: e2e tests

* [ ci skip ]  Hide ROI calculator

So hide until we have time to update it

* Improve: Decrease button area

Before clickable area was expanding to fill entire thing. In the full page view mode this took up a lot of space. Now is expected icon size.

* Clean: Old comments

* Improve: Use header to enable expansion usage

-> This way the show/hide of file level attributes is fully enabled
-> Remove open becuase it was causing an error, whenever another accordion was opened, it would pop open again. As far as I can tell it works fine without it (once opened it stays open for users session), and if we do want it then need a deeper implementation ( was a console error too).

* Improve: Move attributes into accordion and enable for instances

-> Follows pattern from new global attributes
-> Enable existing panel for annotations by add header in default case.
-> Move counts into header, more we can probably do here in future

* clean

* Add: Missing variable relevant in new accordion case

* Add: Icons and fix spacing

* Improve: move to standard expansion panel

-> Not sure why was using a manual open /close before
-> This aligns styles better
-> Including click box aligning with the other items

* Wordsmith: Adjust explanation

-> Also restore the "What Labels Are Attached" to " When Do You Want to Show This?" .  guessing a merge conflict

* fix: styles

* Clean: Logs and formatting [ ci skip ]

* [ci skip]  Add: Only show if there are global attributes

* Restore: Add files to query

Unless this is breaking something - but I think it's nice to have

* fix: unit tests

* fix: e2e tests

* fix: e2e tests

* fix: reduce parallel runners

* fix: tests

Co-authored-by: Pablo <pjestradac@gmail.com>
Co-authored-by: PJEstrada <pjestrada006@hotmail.com>