Improve: Visibility of Issues
1) Make issues according viewable by default. Before had to create an issue to see the existing issues (in this context).
2) Add a "new issue" button. This means a user does not have to know to right click on canvas. Also makes this accessible to other media types more easily
3) Add a "no issues" style art to visual. Fix v-cloack

New var names
show_issue_panel -> show_modify_an_issue  (it's a singlur issue, not a panel)
minimize_issues_sidepanel -> issues_expansion_panel  (it's the v-model, and in expansion now)

Tried to improve formatting in code a bit

The `issues_list` thing does not seem to update very , or pull from server, but that seems like a seperate bug does . These changes are just making it more visible.

Creating an issue still opens the expansion panel, so as far as I can tell there is no loss of that.