Improve: Clean Formatting
Remove extra formatting
Now it fills the space better
The built-in formatting with the card etc still has boundaries
now visually cleaner

Also removed title. This one is more debatable. The justification is: the context it should be obvious what's happening. the title is large and distracting. This focuses more on the users action of actually creating the issue