Make Issues Concept more Visible & Fix global attribute label file permission (#639)
* Fix Saving Issue with Global Attributes

This function probably expected the value to be None.
Added line that's more declarative.

Future would be better if we can validate differently but small thing.

* Fix: this is undefined

Was throwing console error
Now follows pattern same as above with using `component_ctx`

* Fix: Remove old code

This code was from before moving to the expansion panel. It was causing unexpected behavior like the header disappearing.

An issue with moving this to expansion panels is now the component seems to reset on file change. So may have to think about decoupling that somehow

* Improve: Visibility of Issues

1) Make issues according viewable by default. Before had to create an issue to see the existing issues (in this context).
2) Add a "new issue" button. This means a user does not have to know to right click on canvas. Also makes this accessible to other media types more easily
3) Add a "no issues" style art to visual. Fix v-cloack

New var names
show_issue_panel -> show_modify_an_issue  (it's a singlur issue, not a panel)
minimize_issues_sidepanel -> issues_expansion_panel  (it's the v-model, and in expansion now)

Tried to improve formatting in code a bit

The `issues_list` thing does not seem to update very , or pull from server, but that seems like a seperate bug does . These changes are just making it more visible.

Creating an issue still opens the expansion panel, so as far as I can tell there is no loss of that.

* Improve: Clean Formatting

Remove extra formatting
Now it fills the space better
The built-in formatting with the card etc still has boundaries
now visually cleaner

Also removed title. This one is more debatable. The justification is: the context it should be obvious what's happening. the title is large and distracting. This focuses more on the users action of actually creating the issue

* Improve: Hide if open

* Improve: Remove some not needed formatting

* Improve: hide if no issue

Context of other changes, before but chance this wasn't visible but otherwise was showing the attachment icon