fix: remove old marketing toolbar and update redirects to landing page
When visting "/" route now when the installation is not open source we will redirect to landing page,
If the installation is open source we redirect to login page or dashboard

Also removed the old marketing toolbar buttons and changed some of the menu buttons to go to the landing pages instead of marketing components.

I still haven't deleted the components, but if you think now is the time to remove them I can do it in this PR too.