fix: display of examinations
Issue is that when a super admin applies to the exam. He must likely does not belong to the project so the member data is not avaiaible on the vuex store. For now we default to just show the row with no data. But we might need to restrict applications of users only belonging to project or, create an endpoint to get a project member.