Improve Invite Experience & Bug Fixes (#658)
* fix: remove old marketing toolbar and update redirects to landing page

When visting "/" route now when the installation is not open source we will redirect to landing page,
If the installation is open source we redirect to login page or dashboard

Also removed the old marketing toolbar buttons and changed some of the menu buttons to go to the landing pages instead of marketing components.

I still haven't deleted the components, but if you think now is the time to remove them I can do it in this PR too.

* fix: validate URL Base ends with "/" 

Now we check on startup that URL BASE end with a / to avoid misconstructed links

* fix: fix invite to project flow for annotators

Now project is set automatically and is redirected to /me section

* fix: resume disabled state

* fix: display of examinations

Issue is that when a super admin applies to the exam. He must likely does not belong to the project so the member data is not avaiaible on the vuex store. For now we default to just show the row with no data. But we might need to restrict applications of users only belonging to project or, create an endpoint to get a project member.

Co-authored-by: Pablo <pjestradac@gmail.com>