Improve/Fix: Where the plan comes from
Before, if self.user was provided, then execution would never get to project based plan. This was confusing becuase it meant the project could have a plan associated, but the limit checker would still fail to get it.

In the new approach, the project plan takes priority. I think that makes the most sense business level wise and technical wise. If that's wrong it at least provides clear order of priority we can review in future.

This also happens to reduce duplicate code / potential error, since now if the free plan is created it attaches to both project and user in same block.