Add: Error visual, better success case conditions
Context that we were sometimes showing a success when it was not a success.
Now we must expressly `declare_success`.
As as further check it still has the !is_actively_sending becuase dropzone case flags are a bit odd with "multi-sucesss"

Remove `success_upload` becuase we no longer assume that closing it = success!