Better query access, better discovery UI/UX (#779)
* Make Query Concept more clear

1) We have a lot of query functions but before it was really hidden
New icons make Schema, Studio, and Query concepts more clear
2) `show_explorer_full_screen` it's really bad it needs all three from L183 to L185 to handle case if you are already in the studio, and then click query, of pulling up the bottom sheet.

We still get redudent nav thing but it pulls up sheet at least even if the user is already on the studio

* cast to plural

* Fix: Ignore soft delete by default

* Improve ease of use

before it was bouncing around every time click which was hard on eyes
now in this iteration we disable some of the buttons while loading, and attach the linear loader. still needs work but at least it feels less bouncy. visual position of objects is more generally maintained

* Add: better UI exp

1) add None found concept
2) Clear/reset file list on new query

* Add: Query and view annotations from Schema

* Add: Fancier no results screen

* new design idea

* Improve: no results screen

1) Seperate no results from end of results using page number
2) Fix loading using min width as Pablo said

* Tiny fix