Nuke default_sequences_to_single_frame
This was an old concept we are not using.
Trying to define it at label level was just not a very good way to do it
and it mixed a bunch of concepts. was an early mistake
if we wanted similar functions should think about is a simple UI option, and can save it to UI schema or something like that
It was never even really needed becuase it solved wrong problem with video assumptions.