Add dedicated change button
1) The part that was confusing was that change project was within the project menu.
So it meant that various conditions that made sense for most of the menu did not make sense for change project.
2) Now there is a dedicated change project, that will only show up when there is no active project, the user's email is verified, and there is a project available
3) Now the project/task menu only shows when there's an active project.
While there's some UI trade offs for now that seems simpler then the disabled
4) Note for reactivity and context it seems that we need it in the menu and button

while we could "get smart" about some of the other resume onboarding, change, new project etc. there are so many contextual things for now feels better to keep that screen open - again keeping in mind it's a fallback that users should rarely get to