Improve defensive aspects, schema context, and logging
-> Add defensive iterator to walk over labelbox random errors
For example we see Error 520 Labelbox Web server is returning an unknown error
"There is an unknown connection issue between Cloudflare and the origin web server. As a result , the web page can not be displayed."

Now it will store the exception and exception count on the log and keep trying to iterate.

-> Better use http log to return migration results to end user (more useful then server logs in this context)

-> If a label already exists it is skipped. However, that label may be in a different Schema so now we attach it to the new schema if the label had already existed