Solve Fernet issue by storing without the `b`
This is to solve ValueError: Fernet key must be 32 url-safe base64-encoded bytes.

`base64.urlsafe_b64encode(os.urandom(32))` returns bytes
Fernet() can take either if stored in memory

But in this context we are writing to a file so this gets stored as `b'keyliteral'`
When loaded by os.environ.get() it then gets formatted as
`"b'keyliteral'"` (note extra quotation marks).

This was not happening before, it's not clear why in some cases it happens.
There does not seem to be a good way to go from `"b'keyliteral'"` -> `b'keyliteral'` or `'keyliteral'` after the fact, hence having to do it here.

At the moment this issue has only been observed in the docker compose context, not in helm/k8s/bare metal

Tested a number of things like new versions of crypto etc

For future reference we are using `os.urandom(32)` because it stores better then Fernet.generate_key(). And as far as I can tell the issue was with the env loading, not with  `os.urandom(32)`