Prevent confusing state
It was possible before to have project.ABC and project.current.XYZ
Given that we have not yet developed the `project` object to really be anything more then the project.current it's confusing. So at least now `current` doesn't have to be called.
if we don't want this, then we should just delete project.ABC BUT that would require checking a lot of code so for now this.

Also clean old comments