Improve user experience
Prior there was no final block. When first added final block issue was that there are integrated for loops which call more things. That meant that the first loop would exit and it would declare success early. Now even if there are 1000 attributes it only declares success when actually done