Improve auto schema creation (#1015)
* Add default schema support

* Clean duplicate line

* not used anywhere

* Null handling

`label_file_id` key exists but is null

* Add getting default if not provided

* Add tree view support WIP

* Update AttributeGroupManager.ts

* fix: export tree view generation

* cleanup

* fix: global instance count

* fix: concurrency

* Shift to final block to prevent user error

1) Shift to final, otherwise with large schema user may click button multiple times/ may reload, then it throws error again etc becuase it's still "missing" etc
2) opinon shift "error" to info since common expected case is schema creation which is info not error. future can have two types etc

* Improve user experience

Prior there was no final block. When first added final block issue was that there are integrated for loops which call more things. That meant that the first loop would exit and it would declare success early. Now even if there are 1000 attributes it only declares success when actually done

* Add attached annotations filter

Useful when sending a lot of data and want to see something like "Success" + "has attachments"

* fix: tree view options mapping

* bugfix UI

Still bounced around on UI, this zaps it

* cleanup

Co-authored-by: Pablo <pjestradac@gmail.com>