Clarify control flow
So __determine_entity_type() was formatting it. then later when we did generate_query_element() we assumed if it wasn't part of the query we would just return the token.value.

Where this becomes confusing is that we call
                value_1 = self.__parse_value(name1)
                value_2 = self.__parse_value(name2)
BUT when we look at generate_query_element() it looks like it's always returning a relevant entity

So to improve this
1) __determine_entity_type -> __format_entity()
2) add determine_if_reserved_word() to make it super clear where the control flow is.

Maybe overtime the idea was to use something like ListQueryElement() but at least at this moment clarifying the reserved classes feels like the bigger priority

Ideally I'd actually like a function like `determine_if_scaler_or_reserved()` but working within current parsing thing this was closest