fix: performance fixes for projects with large dirs list (#1116)
* Remove project serialization on endpoints to avoid rendering a full list of dirs.
* Limited the number of dirs to be added to project's cache

Eventually I think the best way to go is to completely remove the cache of dirs in project. The only reason for it's existence is for having a dir to get when changing projects. But the correct way should be to have a paginated dir list and then go to the first dir on that paginated list. That way the backend will not serialize huge jsons when there are 3K dirs on a project.

Other thing that was removed is the user visit history component. The query is now extremly slow due to the event table having no indexes. We cannot add indexes to this table since it is very write intensive. For now I'll remove it until we have the time to define a better way to store and query events.