Refactor: disambiguate between values and values metdata
Example error, on item of interest file, and group by date, the `values` array has data, not the `values_metadata`.

So, refactored `append_values()` to be generic to allow to disambiguate between them.

At the moment it tries to get it from this.values then fails over to this.values_metadata. That grouping of logic will likely change, but at least the assumptions around that are surfaced now.

Also hardening null checks since generally assume there is at least one element present