Create Report.ts
Besides general benefits of stronger type, this is to disambiguate between `labels` in different contexts. For example there is report (e.g. charting) labels. And there is Schema Labels.  This uses the new Schema type. So this way instead of having to guess if label is one or the other, it's clear it's Schema.label