Remove duplicate code that was preventing flexible scaling
This code confused concerns. For example, we may wish to do something totally different with a user object once we have an ID. Or formatting may be done on the front end etc.
Second, we had  some of this moved into build_user_metadata() already, which is more composable
Third, with more complex reports, the expected form of the groupings is less clear, so it's better to make less assumptions about re-shaping the data here.