update to new list by tuples format
'list_tuples_by_period' could be just 'list_tuples' so it's more standard (for non date types)

The main idea here is that labels is defined, and then everything else maps to that
so it should never be unclear the relation between labels and values

prior, if we did not pre-format the data by filling zeroes it was unclear
and then if we did fill zeroes, how and when we wrote that data may still be unclear, e.g. since a list full of zeroes makes sense for one format, but then is undefined for another etc. basically it starts coupling the code together, but this should be able to run straight from backend without any further formatting