Add TaskTimeTracking
With the new `init_query` class it's fairly easy to define a different function for the case, e.g. `init_query.group_by = func.sum(self.base_class.time_spent)`

Clarify `time_created_normalized` is the normalized, and add exception if not defined

Over time, we can move some of these if statements to classes for the type, e.g. for TaskTimeTracking some of that can be in the class itself

This way we get a whole bunch of new group by stuff, we can keep formatting changes etc. more in line etc.